requirements.txt
streamlit
google-genai
python-dotenv
pypdf
reportlab
langsmith
Pillow